requirements.txt
pyYaml
requests
confluent_kafka
